Skip packages with compilation errors
Also updatesfixtures that relied on compile errors being ignored
