+++ compiler-plugin-test/src/test/resources/test-src/package_05/types.bal
@@ -1,0 +1,40 @@
+// Copyright (c) 2025 WSO2 LLC. (http://www.wso2.com).
+//
+// WSO2 LLC. licenses this file to you under the Apache License,
+// Version 2.0 (the "License"); you may not use this file except
+// in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+// Hash result containing the hashed value and algorithm used
+public type HashResult record {|
+    string value;
+    string algorithm;
+    string originalText;
+    json...;
+|};
+
+// Base64 encoding/decoding result
+public type Base64Result record {|
+    string result;
+    string operation; // "encode" | "decode"
+    string originalInput;
+    json...;
+|};
+
+// Supported hash algorithms
+public enum HashAlgorithm {
+    MD5 = "md5",
+    SHA1 = "sha1",
+    SHA256 = "sha256",
+    SHA384 = "sha384",
+    SHA512 = "sha512"
+}
